Support custom claims for header injection
In the existing configuration we can inject headers in the user's
response with the value of claims we find in the id token. Currently
this feature supports only a limited set of claims.

With this change we extend the support to get the value from arbitrary
claims and inject them to the user provided header.

Refs #631
Refs #772
Refs #2193
Refs #2757

Signed-off-by: Marjus Cako <marios.cako@hpe.com>